Move peers-specific code to own module
