+++ ouroboros-consensus-diffusion/test/consensus-test/Test/Consensus/PointSchedule.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE DeriveGeneric         #-}
 {-# LANGUAGE DuplicateRecordFields #-}
 {-# LANGUAGE FlexibleContexts      #-}
 {-# LANGUAGE LambdaCase            #-}
@@ -32,10 +31,7 @@ module Test.Consensus.PointSchedule (
   , GenesisWindow (..)
   , HeaderPoint (..)
   , NodeState (..)
-  , Peer (..)
-  , PeerId (..)
   , PeerSchedule
-  , Peers (..)
   , PointSchedule (..)
   , PointScheduleConfig (..)
   , TestFrag
@@ -48,8 +44,6 @@ module Test.Consensus.PointSchedule (
   , fromSchedulePoints
   , genesisAdvertisedPoints
   , longRangeAttack
-  , mkPeers
-  , peersOnlyHonest
   , pointSchedulePeers
   , prettyGenesisTest
   , prettyPointSchedule
@@ -59,17 +53,13 @@ module Test.Consensus.PointSchedule (
 
 import           Control.Monad.ST (ST)
 import           Data.Foldable (toList)
-import           Data.Hashable (Hashable)
 import           Data.List (mapAccumL, partition, scanl', transpose)
-import           Data.List.NonEmpty (NonEmpty ((:|)))
+import           Data.List.NonEmpty (NonEmpty)
 import qualified Data.List.NonEmpty as NonEmpty
-import           Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import           Data.Maybe (fromMaybe, listToMaybe)
-import           Data.String (IsString (fromString))
 import           Data.Time (DiffTime)
 import           Data.Word (Word64)
-import           GHC.Generics (Generic)
 import           Ouroboros.Consensus.Block.Abstract (WithOrigin (..), getHeader)
 import           Ouroboros.Consensus.Protocol.Abstract (SecurityParam,
                      maxRollbacks)
@@ -83,6 +73,8 @@ import qualified System.Random.Stateful as Random
 import           System.Random.Stateful (STGenM, StatefulGen, runSTGen_)
 import           Test.Consensus.BlockTree (BlockTree (..), BlockTreeBranch (..),
                      prettyBlockTree)
+import           Test.Consensus.PointSchedule.Peers (Peer (..), PeerId (..),
+                     Peers (..), getPeerIds, mkPeers, peersList)
 import           Test.Consensus.PointSchedule.SinglePeer
                      (IsTrunk (IsBranch, IsTrunk), PeerScheduleParams (..),
                      SchedulePoint (..), defaultPeerScheduleParams, mergeOn,
@@ -180,62 +172,6 @@ instance Condense NodeState where
     NodeOnline points -> condense points
     NodeOffline -> "*chrrrk* <signal lost>"
 
--- | Identifier used to index maps and specify which peer is active during a tick.
-data PeerId =
-  HonestPeer
-  |
-  PeerId String
-  deriving (Eq, Generic, Show, Ord)
-
-instance IsString PeerId where
-  fromString "honest" = HonestPeer
-  fromString i        = PeerId i
-
-instance Condense PeerId where
-  condense = \case
-    HonestPeer -> "honest"
-    PeerId name -> name
-
-instance Hashable PeerId
-
--- | General-purpose functor associated with a peer.
-data Peer a =
-  Peer {
-    name  :: PeerId,
-    value :: a
-  }
-  deriving (Eq, Show)
-
-instance Functor Peer where
-  fmap f Peer {name, value} = Peer {name, value = f value}
-
-instance Foldable Peer where
-  foldr step z (Peer _ a) = step a z
-
-instance Traversable Peer where
-  sequenceA (Peer name fa) =
-    Peer name <$> fa
-
-instance Condense a => Condense (Peer a) where
-  condense Peer {name, value} = condense name ++ ": " ++ condense value
-
--- | General-purpose functor for a set of peers.
---
--- REVIEW: There is a duplicate entry for the honest peer, here. We should
--- probably either have only the 'Map' or have the keys of the map be 'String'?
---
--- Alternatively, we could just have 'newtype PeerId = PeerId String' with an
--- alias for 'HonestPeer = PeerId "honest"'?
-data Peers a =
-  Peers {
-    honest :: Peer a,
-    others :: Map PeerId (Peer a)
-  }
-  deriving (Eq, Show)
-
-instance Functor Peers where
-  fmap f Peers {honest, others} = Peers {honest = f <$> honest, others = fmap f <$> others}
-
 -- | A tick is an entry in a 'PointSchedule', containing the peer that is
 -- going to change state.
 data Tick =
@@ -260,14 +196,6 @@ tickDefault PointScheduleConfig {pscTickDuration} number active =
 tickDefaults :: PointScheduleConfig -> [Peer NodeState] -> [Tick]
 tickDefaults psc states =
   uncurry (tickDefault psc) <$> zip [0 ..] states
-
--- | A set of peers with only one honest peer carrying the given value.
-peersOnlyHonest :: a -> Peers a
-peersOnlyHonest value =
-  Peers {
-    honest = Peer {name = HonestPeer, value},
-    others = Map.empty
-    }
 
 -- | A point schedule is a series of states for a set of peers.
 --
@@ -307,32 +235,12 @@ defaultPointScheduleConfig =
 -- Accessors
 ----------------------------------------------------------------------------------------------------
 
--- | Extract all 'PeerId's.
-getPeerIds :: Peers a -> NonEmpty PeerId
-getPeerIds peers = HonestPeer :| Map.keys (others peers)
-
 -- | Get the names of the peers involved in this point schedule.
 -- This is the main motivation for requiring the point schedule to be
 -- nonempty, so we don't have to carry around another value for the
 -- 'PeerId's.
 pointSchedulePeers :: PointSchedule -> NonEmpty PeerId
 pointSchedulePeers = peerIds
-
--- | Convert 'Peers' to a list of 'Peer'.
-peersList :: Peers a -> NonEmpty (Peer a)
-peersList Peers {honest, others} =
-  honest :| Map.elems others
-
--- | Construct 'Peers' from values, adding adversary names based on the default schema.
--- A single adversary gets the ID @adversary@, multiple get enumerated as @adversary N@.
-mkPeers :: a -> [a] -> Peers a
-mkPeers h as =
-  Peers (Peer HonestPeer h) (Map.fromList (mkPeer <$> advs as))
-  where
-    mkPeer (pid, a) = (pid, Peer pid a)
-    advs [a] = [("adversary", a)]
-    advs _   = zip enumAdvs as
-    enumAdvs = (\ n -> PeerId ("adversary " ++ show n)) <$> [1 :: Int ..]
 
 ----------------------------------------------------------------------------------------------------
 -- Conversion to 'PointSchedule'
